hashes/requirements.txt
Django==2.2.15
djangorestframework==3.10.3